# Replace the webpack+acorn transform with a Babel plugin