--- a/external/builder/fixtures_esprima/evals-expected.js
+++ b/external/builder/fixtures_esprima/evals-expected.js
@@ -3,11 +3,17 @@ var b = true;
 var c = true;
 var d = false;
 var e = true;
-var f = 'text';
+var f = "text";
 var g = {
- "obj": { "i": 1 },
- "j": 2
+  obj: {
+    i: 1
+  },
+  j: 2
 };
-var h = { 'test': 'test' };
+var h = {
+  test: "test"
+};
 var i = '0';
-var j = { "i": 1 };
+var j = {
+  i: 1
+};