--- a/external/builder/fixtures_esprima/evals.json
+++ b/external/builder/fixtures_esprima/evals.json
@@ -1,1 +1,1 @@
-{ 'test': 'test' }
+{ "test": "test" }