--- a/external/builder/fixtures_esprima/importalias-expected.js
+++ b/external/builder/fixtures_esprima/importalias-expected.js
@@ -1,7 +1,5 @@
-import { Test } from 'import-name';
+import { Test } from "import-name";
 import { Test2 } from './non-alias';
-export {
- Test3
-} from 'import-name';
-var Imp = require('import-name');
+export { Test3 } from "import-name";
+var Imp = require("import-name");
 var Imp2 = require('./non-alias');