… external/builder/fixtures_esprima/blocks-expected.js
… external/builder/fixtures_esprima/comments-expected.js
… external/builder/fixtures_esprima/ifs-expected.js
… external/webpack/pdfjsdev-loader.mjs
… package.json