I was playing around with the PDF.js build process, and I noticed that webpack runs both some Babel plugins, and a custom acorn-based transform (`pdfjsdev-loader`).

I tried to merge the Babel and the Acorn passes, to skip a AST->string->AST round-trip. With the two separate transforms this is what happens:
1. Babel parses the code
2. Babel transforms the AST
3. Babel generates the code
4. Acorn parses the code
5. pdfjsdev-loader transforms the AST
6. `@javascript-obfuscator/escodegen` generates the code
7. Webpack parses the file
8. Webpack concatenates the files

With this PR, this is reduced to:
1. Babel parses the code
2. Babel transforms the AST
3. babel-plugin-pdfjs-preprocessor transforms the AST
4. Babel generates the code
5. Webpack parses the file
6. Webpack concatenates the files

On my machine, this gives a consistent speedup in build times. `gulp lib` went **down from 3.4s to 2.6s** and `gulp dist` went **down from 36s to 29s**. Also, the Babel plugin is smaller than the acorn transforms because Babel provides two utilities that the plugin can heavily rely on:
- `t.valueToNode(val)`, to convert a JS value into the equivalent AST node. This is used when inlining JSON in the code.
- `path.evaluate()`, to statically evaluate simple expressions. This is used when simplifying the various build-time conditions to strip away unused code.

The transform was tested, so we know that the new plugin keeps the old behavior with a couple differences:
- Babel formats code differently from `@javascript-obfuscator/escodegen`, so there are some whitespaces/quotes differences. This doesn't matter in practice, as when running the transform in the full build process and not in isolation Babel was already part of the process, so it was already affecting code formatting.
- In `external/builder/fixtures_esprima/constants-expected.js` there is a `var m = '1' === true;` that is now transformed to `var m = false;`. This is ok because the runtime behavior is the same, and keeping the old behavior would require more work because we cannot use `path.evaluate()` directly.

Also, I didn't implement the `saveComments` option of `pdfjsdev-loader`. That option's only working (and used) value was `false`, because `acorn` already discards all the comments by default so there was never any comment attached to the AST.

I opened this PR because I saw that you already use a custom Babel plugin (`babelPluginReplaceNonWebpackImport`), but it doesn't actually affect users of pdf.js (it only speeds up the build for devs that integrate in in their app), so I won't be offended if you tell me that from a maintenance perspective the current setup is better 😬 However, this PR also has the advantage of removing some deps withuot introducing new ones.