@nicolo-ribaudo 2024-01-22
Oh I forgot to mention that this makes the error messages when using `PDFJSDev.*` improperly slightly better. For example, for `PDFJSDev.eval("BUNDLE_VERSION FOO")`, the old error is
<img width="1157" alt="image" src="https://github.com/mozilla/pdf.js/assets/7000710/bc5e8681-7837-4215-be12-e24fed63c3e4">
while the new error is
<img width="1316" alt="image" src="https://github.com/mozilla/pdf.js/assets/7000710/ad3355a9-e9a8-4309-8fbd-a24fb149235b">