@nicolo-ribaudo 2024-01-22
(squashing now)

> * Since you only mention `gulp lib`, please keep in mind that e.g. the `gulp generic` and `gulp mozcentral` builds are much more important overall (note that `gulp lib` is only used to allow unit-testing in Node.js).

- `gulp generic` goes from 5.5s to 4.0s
- `gulp mozcentral` goes from 4.7s to 3.2s

> 
> * Have you checked that this passes _all_ tests, by successfully running `gulp test` (or at least `gulp test --noChrome`) locally?
>       See [Wiki: Contributing (4 run lint and testing) ()](https://github.com/mozilla/pdf.js/wiki/Contributing#4-run-lint-and-testing) and note e.g. the "Generating reference images"-section (needs to be done against the `master` branch _before_ testing your patch.)

Yes :) But only with `--noChrome` because Chrome tests for me are failing with this error that I haven't figured out yet how to solve: `Error while starting chrome: Network.enable timed out. Increase the 'protocolTimeout' setting in launch/connect calls for a higher timeout if needed.`.

> * How does this affect the final size of the builds, in particular the output of `gulp generic` _and_ `gulp mozcentral`, when compared to the `master` branch?

- `gulp generic`
  - `pdf.sandbox.mjs`: 612 KiB to 614 KiB (this seems to be due `@javascript-obfuscator/escodegen` being smarter than Babel when deciding aboutsingle vs double quotes for strings to avoid unnecessary escapes in the inlined `PDF_SCRIPTING_JS_SOURCE`)
  - `pdf.mjs`: 597 KiB (no change)
  - `pdf.worker.mjs`: 1.79 MiB (no change)
  - `viewer.mjs`: 510 KiB (no change)
- `gulp mozcentral` has no size change in any output file