@nicolo-ribaudo 2024-01-23
Uhm I'm not sure why I did't catch those differences locally (not even the firefox-related ones). I'll try diffing the output bundles to see what could cause it.