@nicolo-ribaudo 2024-01-23
Or is it expected that sometimes there are some slight variations in color of some pixels? e.g.
<img width="337" alt="image" src="https://github.com/mozilla/pdf.js/assets/7000710/578c1758-f7e2-408b-b877-26f5072e3e5b">

I see that some PRs are merged with this type of failures.